Handle detecting docker bridge interfaces
It seems that if there are any docker network bridge interfaces
present on Linux (Unix?) machine (tested on Ubuntu 17.10), it makes
ifcfg crash while detecting devices due to the different naming
pattern.

As described here:
https://github.com/docker/labs/blob/master/networking/concepts/05-bridge-networks.md
docker creates network bridge interfaces with the following pattern,
e.g.: `br-4bcc22f5e5b9`, which fails to match on any currently given
regexp pattern.

This commit simply adds hypen to the `UnixParser`, `get_patterns`'s
first regexp pattern in order to catch interfaces with hypens as well.

As far as the escaping of the hypen or its position within the pattern
goes, my understanding is that the both cases are "safe" (having it as
the first/last character in the group or escaping it with backlash).

Here's an SO question/answer which realting a very similiar use case:
https://stackoverflow.com/questions/9589074/regex-should-hyphens-be-escaped/9589642#9589642

This change also includes a test for this particular case.