Add support for multiple netmasks, broadcast addresses, as well as ipv6 prefix lengths
This adds `netmasks`, `broadcasts`, and `prefixlens` attributes as lists. For
backwards compatibility, copy the first item in `netmasks` and `broadcasts` to
`netmask` and `broadcast` respectively.

Also, a minor tweak to `tools_test.py`, which fails when run on macOS. (I believe
this is likely due to the ancient version of `bash` that `/bin/sh` is.)